Various cleanups in AdGuard Home
